Can restart Dispatcher during operation
The Dispatcher now runs in a Deployment. When restared, it finds
existing containerized functions and "adopts" them. This is a step
along the way to recompiling the Dispatcher for decontainerization.

Therefore, the Dispatcher no longer deletes the Services and
ReplicaSets for running serverless functions on SIGTERM. Instead,
we have a new Controller that does so. We will end up using the
Controller to compile traces in an upcoming patch.

Finally, we renamed function-storage to just storage. The code we
wrote assumes that the function- prefix is used for serverless
functions.