Address a build issue
The k8s package depends on the kube crate's openapi feature to compile.
However, k8s/Cargo.toml did not specify this requirement. Why did this ever
work? Other packages explicitly depend on kube and specify the openapi feature.
Thus, k8s only builds successfully when it is a depenency of a package that
requires the openapi feature! This is awful.