Faster test running, removed spurious returns
When creating a new function, the "exclusive" flag clears all other
functions. So, we don't have to separately delete and create, which
speeds up testing. Right now, the CLI sets exclusive to true, and
there is no way to unset it from the CLI. But, this is an easy
feature to add.

There were also a number of code blocks in the test_runner that
looked like this:

  assert!(false, ...);
  return results;

I've simplified them.